pyproject.toml
[project]
name = "mcpadapters"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain[openai]>=0.3.25",
    "langchain-mcp-adapters>=0.1.8",
    "langchain-openai>=0.3.23",
    "langgraph>=0.5.0",
    "mcp[cli]>=1.10.1",
    "python-dotenv>=1.1.0",
    "fastmcp>=2.10.2",
    "isort>=6.0.1",
    "black>=25.1.0",
]
